move function into endpoint handler
