+++ cmd/api/middleware.go
@@ -2,11 +2,9 @@ package main
 
 import (
 	"dyno/internal/logger"
-	"dyno/internal/authentication"
 	"net/http"
 	"runtime/debug"
 	"time"
-	"fmt"
 )
 
 func APIGatewayMiddleware(next http.Handler) http.Handler {
@@ -65,17 +63,5 @@ func LoggingMiddleware(next http.Handler) http.Handler {
 			"path", r.URL.EscapedPath(),
 			"duration", time.Since(start),
 		)
-		if r.URL.EscapedPath() == "/login"{
-			var err error
-
-			var code = r.URL.Query().Get("code")
-			var tokenAuthURL = authentication.GetTokenAuthURL(code)
-			var token *authentication.Token	
-			if token, err = authentication.GetToken(tokenAuthURL); err != nil {
-				fmt.Println(err)
-				return
-			}
-			fmt.Printf("%+v",token)
-		}
 	})
 }
