+++ cmd/api/routes.go
@@ -9,7 +9,7 @@ import (
 
 func registerRoutes(service *web.Service, hostdocs bool, hosthealth bool) {
 	endpoints.PostEcho(service)
-	endpoints.GetCode(service)
+	endpoints.Authentication(service)
 	// Swagger UI endpoint at /docs.
 	if hostdocs {
 		service.Docs("/docs", swgui.New)
